Support report_step in merge_rft_ertobs
Extracts in a non-robust way the report_step from the observation filename.
The report_step is not important for the CSV file this script produces, but getting
the well-name right is. create_rft_ertobs.py from fmu-tools creates obs filenames
on this syntax.

Observations returned are now also sorted by well-name++ and reindexed.