Adding ri_wellmod - ResInsight well modelling wrapper (#350)
Adding ri_wellmod - a wrapper script around ResInsight that creates Eclipse well definition keywords based on a ResInsight project (with well setup) and an Eclipse case (initialized or GRDECL input for grid and PERMX|Y|Z + NTG). LGRs are supported, and well LGRs may specified be on the command line.