eclcompress - compress only allowlist keywords by default
The allowlist was generated by parsing the Eclipse reference manual
for keywords that allow repeated counts. It can now compress keywords
within a multi-keyword context (e.g. `EQUALS`) as well as keywords
whose contents and terminator are on the same line.

It defaults to the allowlist but still allows a regex to be applied.
This removes the old way of functioning in which keywords were
evaluated only by a regex which could in some cases apply compression
to keywords which do not allow repeat counts and thereby damage the
file.