Use memoryconservative mode for WF CSV_MERGE
When run as a workflow job/ert plugin this has been observed through
logging to fail due to lack of memory.

When run as a workflow job, it is probable that there is one CSV for
every realization that is to be loaded, thus the memory conservative
option probably makes sense to have on by default.

Keeping the faster option as default for usages as a forward model step.